Add libblockdev-crypto3 as alternative to libblockdev-crypto2
libblockdev-crypto3 replaces libblockdev-crypto2 in Debian
Trixie and higher. Debian Bookworm pulls in libblockdev-crypto2
via a dependency by udisks2; in Trixie, udisks2 depends on
libblockdev-crypto3 instead.

This fixes issues when distro-morphing Debian Trixie and higher
to Kicksecure. It is also necessary for installing Kicksecure on
ports-only architectures such as ppc64.

Incidentally, Trixie replaces Bookworm's "rec" with a "dep", so
it's conceivable that Kicksecure could drop the explicit
dependency once Bookworm support is dropped.